+++ Microsoft.DotNet.Try.Client/src/components/MsDocsCodeRunner.tsx
@@ -9,21 +9,11 @@ import GistPanel from "./GistPanel";
 import TryDotnetBanner from "./TryDotnetBanner";
 import { editorDiv, tryDotnetBanner } from "../dot.net.style.app.css";
 import Frame from "./Frame";
-import { IFrameWindow } from "./IFrameWindow";
+import { GetBlazorIFrameUrl } from "./GetBlazorIFrameUrl";
 
 const MsDocsCodeRunner = (trydotnetWindow: Window) => {
 
-  let trydotnetIframe = new IFrameWindow(trydotnetWindow);
-  let trydotnetHostOrigin = new URL(trydotnetWindow.document.URL).origin;
-  let workspaceType = trydotnetIframe.getQuery().get("workspaceType") || "blazor-console";
-  let clientParams = trydotnetIframe.getClientParameters();
-  if (clientParams && clientParams.workspaceType) {
-    workspaceType = clientParams.workspaceType;
-  }
-  let localRunnerUrl = `/LocalCodeRunner/${workspaceType}/`;
-  let url = new URL(localRunnerUrl, trydotnetHostOrigin);
-  url.searchParams.append("embeddingHostOrigin", trydotnetIframe.getHostOrigin().origin);
-  url.searchParams.append("trydotnetHostOrigin", trydotnetHostOrigin);
+  let url = GetBlazorIFrameUrl(trydotnetWindow);
 
   // @ts-ignore Some connect type inference thing
   let hack = <Frame src={url}/>;
